fix fieldLabel prop type import
